mfg_tool: append common name from DAC to the cn_dac csv file
Earlier it used to append the randome uuid even if common name is
different. If we try to register this csv file with cloud it would fail
validating the cn in DAC and cn specified in the csv file.

So, extract the cn from the subject of the DAC and append it to the
cn_dac csv file

also make sure we match the directory name to common name from the
subject of dac if its of type uuid4, else keep generate the random one

Added tests to validate the above changes